+++ src/main/java/gui/inicioSesion/InicioSesionVista.java
@@ -2,7 +2,6 @@ package gui.inicioSesion;
 
 import controller.TodosController;
 import gui.admin.VistaAdmin;
-import gui.chef.PruebaVistaChef;
 import gui.chef.VistaChef;
 import gui.reponedor.VistaReponedor;
 import inventoryChef.Admin;
@@ -94,7 +93,7 @@ public class InicioSesionVista extends JFrame {
                 if (controller.autenticarRol(controller.validarInicio(username, password)).equals("Chef")) {
                     JOptionPane.showMessageDialog(InicioSesionVista.this, "Inicio de sesión exitoso");
                     SwingUtilities.invokeLater(() -> {
-                        PruebaVistaChef inventarioChef = new PruebaVistaChef((Chef) controller.validarInicio(username, password));
+                        VistaChef inventarioChef = new VistaChef((Chef) controller.validarInicio(username, password));
                         inventarioChef.setVisible(true);
                     });
                     dispose();
